+++ unireg/business/src/main/java/ch/vd/uniregctb/interfaces/service/host/ServiceInfrastructureHostInterfaces.java
@@ -129,7 +129,7 @@ public class ServiceInfrastructureHostInterfaces extends ServiceInfrastructureBa
 	}
 
 	@Override
-	public CommuneId getCommuneIdByEgid(long egid, RegDate date, Long hintNoOfsCommune) throws InfrastructureException {
+	public CommuneId getCommuneIdByEgid(int egid, RegDate date, Integer hintNoOfsCommune) throws InfrastructureException {
 		throw new NotImplementedException("La méthode 'getCommuneByEgid' ne doit pas être appelée sur le service host-interfaces.");
 	}
